Made auto saving/loading fully functional
With this, autoloading upon launching the game should be fully working.
Also added some randomness to the pitch of step_sfx to make them sound a
bit less repetitive, and made it so the game can only save if the
current scene is a Scene_Map (because auto_saving at the titlescreen
breaks things pretty badly).