Fixing bug with window prompts and adding credits message box
The game would try to catch up in frames when a window prompt was
opened, so if the window was open for 140 frames the game would process
the next 140 frames instantly after the game was closed.  To fix this, I
have the window prompt time how long it's opened and then I tell the
game to sleep for however many frames it was opened for.

Credits message works but does not yet have all the desired features,
will finish up later.