Compile with Ruby 2.3, not 2.1 (to fix saves)
I found out what the actual issue was with saves — it wasn’t that
xScripts needed to be recompiled (although it helped load previous
saves from the Windows build on Steam), it was that Ruby 2.1 doesn’t
seem to work with the saves, @mathewv might have determined this when
selecting 2.2 as the minimum.  I have Ruby 2.3 on my system, installed
through Homebrew, so that is what I selected for my particular build.