Added save backup system to the game
To better deal with the random savefile corruptions that occur for some
users.  This should help?  I tested it with each file in the chain
corrupted to make sure it was falling back onto the backups properly,
and made it delete the base file if all backups were corrupt, which is a
very unlikely scenario but might as well put a case to handle it.